Use explicit conversion to uint8_t
The value is already clamped, so it is guaranteed to fit.

Fixes MSVC conversion warning about possible loss of data.

+++ src/States/MapViewStateDraw.cpp
@@ -52,7 +52,7 @@ namespace {
 				glowStepDelta = -glowStepDelta;
 			}
 		}
-		return glowStep;
+		return static_cast<uint8_t>(glowStep);
 	}
 }
